Split StringReader into its own header
