+++ libc/src/stdio/vsscanf.cpp
@@ -11,8 +11,8 @@
 #include "hdr/stdio_macros.h"
 #include "src/__support/CPP/limits.h"
 #include "src/__support/arg_list.h"
-#include "src/stdio/scanf_core/reader.h"
 #include "src/stdio/scanf_core/scanf_main.h"
+#include "src/stdio/scanf_core/string_reader.h"
 
 #include <stdarg.h>
