+++ libc/test/src/stdio/scanf_core/converter_test.cpp
@@ -9,7 +9,7 @@
 #include "src/__support/CPP/string_view.h"
 #include "src/stdio/scanf_core/converter.h"
 #include "src/stdio/scanf_core/core_structs.h"
-#include "src/stdio/scanf_core/reader.h"
+#include "src/stdio/scanf_core/string_reader.h"
 
 #include "test/UnitTest/Test.h"
